// backend/src/maps/types/objectInfo.type.ts
export interface OjbectInfo {
  Name: string;
  NameEng?: string;
  Active?: string;
  ShortDescription?: string;
  ShortDescriptionEng?: string;
  Description?: string;
  DescriptionEng?: string;
  Address: string;
  AddressEng?: string;
  ObjectPhone?: string;
  Email?: string;
  URL?: string;
  WorkInMOFR?: string;
  WorkInSat?: string;
  WorkInSun?: string;
  Area?: any;
  SportComplexType?: string;
}